feat: Add image resizer component and supporting utility functions for image manipulation.
